♻️ Refactor: "jpa방식으로 리팩토링"
Ref : #96
