+++ src/main/java/alcoholboot/toastit/feature/api/dto/LatXLngY.java
@@ -2,6 +2,7 @@ package alcoholboot.toastit.feature.api.dto;
 
 import lombok.Data;
 
+
 @Data
 public class LatXLngY {
     public double lat;
