--- src/main/resources/templates/feature/api/mappers/WeatherMapper.xml
@@ -1,87 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="alcoholboot.toastit.feature.api.repository.WeatherMapper">
-    <select id="selectArea" resultType="AreaRequestDTO" parameterType="Map">
-        <choose>
-            <when test="type == 'city'">
-                SELECT
-                areacode, step1
-                FROM
-                toastitdb.weather_area
-                WHERE
-                step2 is null AND step3 is null
-                ORDER BY
-                step1
-            </when>
-
-            <when test="type == 'county'">
-                SELECT
-                areacode, step2
-                FROM
-                toastitdb.weather_area
-                WHERE
-                step1 = #{keyword} AND step3 is null AND step2 is not null
-                ORDER BY
-                step2
-            </when>
-
-            <when test="type == 'town'">
-                SELECT
-                areacode, step3
-                FROM
-                toastitdb.weather_area
-                WHERE
-                step2 = #{keyword} AND step3 is not null
-                ORDER BY
-                step3
-            </when>
-        </choose>
-    </select>
-
-    <select id="selectCoordinate" parameterType="Map" resultType="AreaRequestDTO">
-        SELECT areacode
-        FROM toastitdb.weather_area
-        WHERE areacode = #{areacode}
-    </select>
-
-    <select id="selectSameCoordinateWeatherList" parameterType="AreaRequestDTO"
-            resultType="WeatherDTO">
-        SELECT DISTINCT baseDate,
-                        baseTime,
-                        category,
-                        nx,
-                        ny,
-                        obsrValue
-        FROM toastitdb.weather_response
-        WHERE baseDate = #{baseDate}
-          AND baseTime = #{baseTime}
-          AND nx = #{nx}
-          AND ny = #{ny}
-    </select>
-
-    <select id="selectAreaCode" parameterType="Map" resultType="String">
-        SELECT areacode
-        FROM toastitdb.weather_area
-        WHERE gridX = #{nx} and gridY = #{ny}
-    </select>
-
-    <insert id="insertWeatherList" parameterType="WeatherDTO">
-        INSERT INTO toastitdb.weather_response(
-        baseDate
-        ,baseTime
-        ,category
-        ,nx
-        ,ny
-        ,obsrValue
-        )
-        VALUES
-        <foreach collection="list" item="item" open="(" separator="),(" close=")">
-            #{item.baseDate}
-            ,#{item.baseTime}
-            ,#{item.category}
-            ,#{item.nx}
-            ,#{item.ny}
-            ,#{item.obsrValue}
-        </foreach>
-    </insert>
-</mapper>
